factored helpers into MVC helpers, and removed some unused imports from mth5_viewer_v1.py, issue #4
